Move placeholder client to new client/subscriptions.py module.
Also a fix for #574

Resolves #494
